Cleaned up leftover merge markers
As I understand it, Git automatically inserts markers like:
 <<<<<<< HEAD
[file content here]
========
[more file content here]

to flag file contents that may need manual merging or when a previous merge may have failed or is incomplete. Probably was my fault as I am still learning the IDE to Git workflow. Anyhow, there was no need for these markers and I have manually edited this file and believe it to be correct.