Add two new classes, two methods
I made two new classes the model FileModel class and the FileIndex
class. I added the method addFile to it which will be called in the
FileIndex class. In the FileIndex class added the addFileToIndex()
method and so far it will open file chooser and restricts users to .txt
files and .html files.