Add methods, changed view
I added new methods in the model class and the index class. I needed to
change the view so that the number of files indexed would use a label
instead of a int so that I can update the label. This caused me to use
another hbox in the main view and a gridpane and hbox in the maintenance
view. In the addFileToIndex method will now allow users to open files
with only files ending in .txt and .html which will save the path of the
file in a list.