Add new class, change view again
I added a new class called Files which is basically the class that will
get the values for the table view. In model I added the code to add the
file and its path into the table view. I also needed to update the
maintenance view code so that the data will show in the table view.