Add stub methods, removeFiles working
I added the stub search option methods to the FileIndex class. I was
able to the get the removeFilesFromIndex method to work and added it to
the remove files button. I updated the removeFiles method to update the
number of files indexed. I also added a checker to the fileChooser to
check if the user hit the cancel button and to just return if they did.