Added inverted index basic build and save code
The BASIC inverted index construction and saving is now in
FileIndex.java. This is very much still a work in progress and I will be
adding and modifying a lot over the next few days. Also the
Test_Sample*.txt files have been placed in the resources folder. The
inverted index JSON file will also be saved to this location.