addToInvIndex now working through GUI Maint. window
As files are selected and added to the Files List, they will also be
added to the Inverted Index Map.

Note that invIndex.json is also saved, temporarily, by the addToInvIndex
method. The saveIndexToStorage will need to be called other ways (like
on program exit) in the future.

TODO:
removeFromInvIndex
updateIndex
loadIndexFromStorage