Added do___Search methods to IndexModel.
Per my convo. with Pollock last night, we should have these (even if
just a placeholder to be removed/replaced later) in our Model code
even if the Controller code already has methods of these types. That
was my understanding of his expectations for this project, so I have
added them. We can further discuss how the ___Search functions will be
codedwhen moving on to the next project submittal.