Add code to both initialize methods, and fixed exception
I added code to the loadIndexFromStorage method and the initializeIndex
method so that it will create the files if they exist when trying to be
loaded and won't throw an exeption. I also fixed a nullpointer exception
that was happening when the application was launched if the user just
left it on and search and tried to do a search.